+++ dags/constants.py
@@ -1,4 +1,5 @@
 rabbitmq_queue_minio_event='afload'
+rabbitmq_queue_object_event='afobjectload'
 rabbitmq_queue_minio_register='afregister'
 rabbitmq_exchange_notify='notify'
 rabbitmq_exchange_notify_key_s3file='s3'
