+++ docker-compose/airflow/docker-compose.yml
@@ -91,6 +91,7 @@ services:
 
   redis:
     image: redis:latest
+    container_name: redis
     ports:
       - 6379:6379
     healthcheck:
@@ -148,6 +149,17 @@ services:
       timeout: 10s
       retries: 5
     restart: always
+
+  redis-commander:
+    container_name: redis-commander
+    hostname: redis-commander
+    image: rediscommander/redis-commander:latest
+#    build: .
+    restart: always
+    environment:
+    - REDIS_HOSTS=local:redis:6379
+    ports:
+    - 8081:8081
 
 volumes:
   postgres-db-volume:
