+++ docker-compose/minio-trino/rabbitmq/definitions.json
@@ -52,6 +52,15 @@
          }
       },
       {
+         "name":"afobjectload",
+         "vhost":"/",
+         "durable":true,
+         "auto_delete":false,
+         "arguments":{
+            
+         }
+      },
+      {
          "name":"afregister",
          "vhost":"/",
          "durable":true,
